+++ src/app/api/batch/backfill-missing-xp/route.ts
@@ -1,7 +1,7 @@
-import { grantXpBatch } from "@/lib/services/userLevel";
+import { grantXpBatch } from "@/features/user-level/services/level";
+import { calculateMissionXp } from "@/features/user-level/utils/level-calculator";
 import { createAdminClient } from "@/lib/supabase/adminClient";
 import { executeChunkedQuery } from "@/lib/utils/supabase-utils";
-import { calculateMissionXp } from "@/lib/utils/utils";
 import { type NextRequest, NextResponse } from "next/server";
 
 // 型定義を明示（Supabaseの!innerJOINが配列を返す可能性を考慮）
