+++ src/app/seasons/[slug]/users/[userId]/page.tsx
@@ -14,13 +14,13 @@
  * - 全シーズン履歴
  */
 import UserDetailActivities from "@/app/users/[id]/user-detail-activities";
-import Levels from "@/components/levels";
 import { Card } from "@/components/ui/card";
 import { SocialBadge } from "@/components/ui/social-badge";
 import { UserMissionAchievements } from "@/components/user-mission-achievements";
 import { UserSeasonHeader } from "@/components/user-season-header";
 import { UserSeasonHistory } from "@/components/user-season-history";
 import { UserBadges } from "@/features/user-badges/components/user-badges";
+import Levels from "@/features/user-level/components/levels";
 import {
   getUserActivityTimeline,
   getUserActivityTimelineCount,
