+++ src/components/ui/progress-bar-simple.tsx
@@ -1,6 +1,10 @@
 "use client";
 
-import { cn, getLevelProgress, getXpToNextLevel } from "@/lib/utils/utils";
+import {
+  getLevelProgress,
+  getXpToNextLevel,
+} from "@/features/user-level/utils/level-calculator";
+import { cn } from "@/lib/utils/utils";
 import React from "react";
 
 interface ProgressBarSimpleProps {
