+++ src/features/onboarding/components/onboarding-mission-card.tsx
@@ -4,8 +4,8 @@ import MissionAchievementStatus from "@/components/mission/mission-achievement-s
 import { Button } from "@/components/ui/button";
 import { Card, CardFooter, CardHeader, CardTitle } from "@/components/ui/card";
 import { MissionIcon } from "@/components/ui/mission-icon";
+import { calculateMissionXp } from "@/features/user-level/utils/level-calculator";
 import type { Tables } from "@/lib/types/supabase";
-import { calculateMissionXp } from "@/lib/utils/utils";
 import clsx from "clsx";
 import { motion } from "framer-motion";
 import { UsersRound } from "lucide-react";
