+++ src/features/user-level/components/levels.test.tsx
@@ -18,7 +18,7 @@ describe("Levels", () => {
         "test-user-id",
       );
       expect(
-        require("@/lib/services/userLevel").getUserLevel,
+        require("@/features/user-level/services/level").getUserLevel,
       ).toHaveBeenCalledWith("test-user-id", undefined);
     });
   });
