+++ src/features/user-level/utils/level-calculator.ts
@@ -1,0 +1,70 @@
+// L → L+1 の差分 XP
+export const xpDelta = (L: number) => {
+  if (L < 1) throw new Error("Level must be at least 1");
+  return 40 + 15 * (L - 1);
+};
+
+// レベル L 到達までの累計 XP
+export const totalXp = (L: number) => {
+  if (L < 1) throw new Error("Level must be at least 1");
+  return (L - 1) * (25 + (15 / 2) * L);
+};
+
+/**
+ * XPに基づくレベル計算
+ * 新しい式に基づく逆算
+ */
+export function calculateLevel(xp: number): number {
+  if (xp < 0) return 1;
+
+  // 最大レベルを設定（計算の無限ループを防ぐため）
+  const maxLevel = 1000;
+
+  for (let level = 1; level <= maxLevel; level++) {
+    const requiredXp = totalXp(level + 1);
+    if (xp < requiredXp) {
+      return level;
+    }
+  }
+
+  return maxLevel;
+}
+
+/**
+ * ミッションの難易度に基づいてXPを計算する
+ */
+export function calculateMissionXp(difficulty: number): number {
+  switch (difficulty) {
+    case 1:
+      return 50; // ★1
+    case 2:
+      return 100; // ★2
+    case 3:
+      return 200; // ★3
+    case 4:
+      return 400; // ★4
+    case 5:
+      return 800; // ★5
+    default:
+      return 50; // デフォルト（Easy相当）
+  }
+}
+
+/**
+ * 次のレベルまでに必要なXP計算
+ */
+export function getXpToNextLevel(currentXp: number): number {
+  const currentLevel = calculateLevel(currentXp);
+  const nextLevelTotalXp = totalXp(currentLevel + 1);
+  return Math.max(0, nextLevelTotalXp - currentXp);
+}
+
+/**
+ * 現在レベルでの進捗率計算（0-1の値）
+ */
+export function getLevelProgress(currentXp: number): number {
+  const currentLevel = calculateLevel(currentXp);
+  const xpToNext = getXpToNextLevel(currentXp);
+  const levelXpRange = xpDelta(currentLevel);
+  return Math.max(0, Math.min(1, (levelXpRange - xpToNext) / levelXpRange));
+}
